release 0.1.6 source code for go

+++ services/cdn/v1/model/model_urls.go
@@ -10,7 +10,7 @@ import (
 type Urls struct {
 
 	// urlid
-	Id *int32 `json:"id,omitempty"`
+	Id *int64 `json:"id,omitempty"`
 
 	// url具体值
 	Url *string `json:"url,omitempty"`
