+++ services/dcs/v2/model/model_resize_instance_body.go
@@ -14,8 +14,8 @@ type ResizeInstanceBody struct {
 	// 产品规格编码。具体查询方法如下：  - 方法一：查询产品介绍中的[实例规格](https://support.huaweicloud.com/productdesc-dcs/dcs-pd-0522002.html) - 方法二：登录分布式缓存的控制台界面，点击购买缓存实例，查找对应的实例规格名称 - 方法三：调用[查询产品规格](https://support.huaweicloud.com/api-dcs/ListFlavors.html)接口查询。
 	SpecCode string `json:"spec_code"`
 
-	// 新的缓存实例规格，新的规格必须大于扩容前的规格，单位：GB。 取值包括：4，8，16，32，64 取值必须是当前产品支持的实例规格，请以实际为准。
-	NewCapacity int32 `json:"new_capacity"`
+	// 新的缓存实例规格，单位：GB。 Redis4.0和Redis5.0：单机和主备类型实例取值：0.125、0.25、0.5、1、2、4、8、16、32、64。Cluster集群实例规格支持24、32、48、64、96、128、192、256、384、512、768、1024。 Memcached：单机和主备类型实例取值：2、4、8、16、32、64。
+	NewCapacity float32 `json:"new_capacity"`
 
 	BssParam *BssParamEntity `json:"bss_param,omitempty"`
